+++ crates/rtlola-hir/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - TBD
+
+### Added
+- Added the `FeatureSelector` to check for unsupported language features.
+- Added `is_eval_filtered` method to stream trait.
+
 ## [0.4.1] - 02.02.2023
 
 ### Changed